104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,14 +10,14 @@ int main(void)
 {
 	int n = 98;
 
-	long long a = 1, b = 2, c, i;
+	long a = 1, b = 2, c, i;
 
-	printf("%lld, %lld", a, b);
+	printf("%ld, %ld", a, b);
 
 	for (i = 3; i <= n; i++)
 	{
 		c = a + b;
-		printf(", %lld", c);
+		printf(", %ld", c);
 		a = b;
 		b = c;
 	}
